move files to their appropriate locations
