+++ garden/src/features/gardens/components/create/CreateGardenFormFields.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { UseFormReturn, useFormContext } from "react-hook-form";
 
 import { ChevronDown, ChevronUp } from "lucide-react";
